Move middleware utils to utils-file
